Cycle and Speed Fix

+++ Assets/Scripts/Entities/BattleEntities/Enemy.cs
@@ -9,7 +9,7 @@ public class Enemy : Character
     public int id;
     public int sp;
 
-    public Enemy(string name, int identifier, int le, int hp, double atk, double def, double spd, double cr, double cd) : base(name,le, hp, atk, def, spd, cr, cd) {
+    public Enemy(string name, int identifier, int le, int hp, double atk, double def, int spd, double cr, double cd) : base(name, le, hp, atk, def, spd, cr, cd) {
         character = name;
         level = le;
         baseHP = hp;
